+++ inst/shiny/modules/tab_nca/setup/slope_selector.R
@@ -135,7 +135,9 @@ slope_selector_server <- function( # nolint
 
       if (changes$in_data) {
         # New data or major changes: regenerate all plots
-        plot_outputs(get_halflife_plots(new_pknca_data)[["plots"]])
+        plot_outputs(get_halflife_plots(
+          new_pknca_data, title_vars = "ATPTREF"
+        )[["plots"]])
       } else if (changes$in_hl_adj) {
         # Modify plots with new half-life adjustments (inclusions/exclusions)
         plot_outputs(handle_hl_adj_change(new_pknca_data, pknca_data(), plot_outputs()))